requirements.txt
keras==2.13.1
numpy==1.24.2
scipy==1.10.1
tensorflow==2.13.0
mxnet==1.6.0
matplotlib==3.7.1
Pillow==9.5.0